% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropout.R
\name{dropout}
\alias{dropout}
\title{Compute dropout rate population from a target population}
\format{
A numeric value of number of dropouts or percentage of dropouts
}
\usage{
dropout(first_dose, last_dose, percentage = TRUE)
}
\arguments{
\item{first_dose}{number that received the first vaccine dose}

\item{last_dose}{number that received the last vaccine dose}

\item{percentage}{whether to return the actual number or percentage}
}
\description{
Compute the population target that who start the immunization schedule but dropped out i.e didn't who complete it.
}
\examples{
\dontrun{
# under dropout of 800 children who received DTP 3 vaccine out of 900 who received DTP 1

dropout(first_dose = 900, last_dose = 800)
}
}
