% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/under_immunized.R
\name{under_immunized}
\alias{under_immunized}
\title{Compute under immunized population from a target population}
\format{
A numeric value of number of immunized population or percentage of immunized population
}
\usage{
under_immunized(n_received, target_population, percentage = FALSE)
}
\arguments{
\item{n_received}{actual number that received the last vaccine dose (DTP 3) or Measles depending on national immunization guideline}

\item{target_population}{target population estimate}

\item{percentage}{whether to return the actual number or percentage}
}
\description{
Compute the population target that didn't receive all the due vaccines as per the national immunization guideline
}
\examples{
\dontrun{
# under immunized population of 800 children who received DTP 3 vaccine out of 1000 population

under_immunized(n_received = 800, target_population = 1000)
}
}
