% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_dose.R
\name{zero_dose}
\alias{zero_dose}
\title{Compute zero dose population}
\format{
A numeric value of number of zero dose or percentage of zero dose
}
\usage{
zero_dose(n_received, target_population, percentage = FALSE)
}
\arguments{
\item{n_received}{actual number that received the vaccine (DTP 1)}

\item{target_population}{target population estimate}

\item{percentage}{whether to return the actual number or percentage}
}
\description{
Compute the population target that didn't receive DTP 1 vaccine
}
\examples{
\dontrun{
# Zero dose of 900 children who received a vaccine out of 1000 population

zero_dose(n_received = 900, target_population = 1000)
}
}
